requirements.txt
asgiref==3.3.1
decorator==4.4.2
decorators==2.0.7
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
GDAL==3.1.3
Pillow==8.1.0
psycopg2-binary==2.8.6
pytz==2020.5
self==2020.12.3
sqlparse==0.4.1
style==1.1.6
update==0.0.1
filters~=1.3.2
requests~=2.25.1
enum34~=1.1.6
typing~=3.7.4.3
six~=1.15.0